# paths.nim scanline method perf improved