… tests/images/svg/Ghostscript_Tiger.png
… tests/images/svg/ellipse01.png
… tests/images/svg/quad01.png
… tests/images/svg/rect02.png