`nim c -d:release --debugger:native -f -r .\tests\benchmark_svg.nim`
before:
`svg decode ........................ 22.328 ms     22.950 ms    ±0.353   x218`
after:
`svg decode ........................ 20.526 ms     21.223 ms    ±0.397   x235`

`nim c -d:release --debugger:native -r .\experiments\benchmark_cairo.nim`
before:
`pixie .............................. 0.250 ms      0.256 ms    ±0.007  x1000`
after:
`pixie .............................. 0.210 ms      0.220 ms    ±0.017  x1000`